Working feign with int test example
