--- src/main/java/uk/gov/hmcts/reform/ccd/feign/FeignConfiguration.java
@@ -1,19 +1,0 @@
-package uk.gov.hmcts.reform.ccd.feign;
-
-import feign.Retryer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class FeignConfiguration {
-
-    @Bean
-    public Retryer feignRetryer() {
-        return new Retryer.Default(100, 1000, 3);
-    }
-
-    @Bean
-    public feign.codec.ErrorDecoder errorDecoder() {
-        return new ErrorDecoder();
-    }
-}
