+++ src/main/java/uk/gov/hmcts/reform/ccd/feign/RetryableFeignConfig.java
@@ -1,0 +1,17 @@
+package uk.gov.hmcts.reform.ccd.feign;
+
+import feign.Retryer;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import uk.gov.hmcts.reform.ccd.parameter.ParameterResolver;
+
+@Configuration
+public class RetryableFeignConfig {
+
+
+    @Bean
+    public Retryer feignRetryer() {
+        // Example: Retry 3 times, waiting 100ms, then 200ms, then 300ms, etc.
+        return new Retryer.Default(100, 200, 3);
+    }
+}
